requirements.txt
underthesea
mysql-connector-python
mysqlclient
requests-html